+++ packages/extension-polkagate/src/components/translate.ts
@@ -1,10 +1,11 @@
 // Copyright 2019-2024 @polkadot/extension-polkagate authors & contributors
 // SPDX-License-Identifier: Apache-2.0
-// @ts-nocheck
 
-import { useTranslation as useTranslationBase, UseTranslationResponse, withTranslation } from 'react-i18next';
+import type { UseTranslationResponse } from 'react-i18next';
 
-export function useTranslation(): UseTranslationResponse<string, undefined> {
+import { useTranslation as useTranslationBase, withTranslation } from 'react-i18next';
+
+export function useTranslation (): UseTranslationResponse<string, undefined> {
   return useTranslationBase();
 }