+++ packages/extension-polkagate/src/util/utils.ts
@@ -16,7 +16,8 @@ import { BN, BN_TEN, BN_ZERO, hexToBn, hexToString, hexToU8a, isHex, stringToU8a
 import { decodeAddress, encodeAddress } from '@polkadot/util-crypto';
 
 import { EXTRA_PRICE_IDS } from './api/getPrices';
-import { ASSET_HUBS, BLOCK_RATE, FLOATING_POINT_DIGIT, INITIAL_RECENT_CHAINS_GENESISHASH, PROFILE_COLORS, RELAY_CHAINS_GENESISHASH, SHORT_ADDRESS_CHARACTERS } from './constants';
+import allChains from './chains';
+import { ASSET_HUBS, BLOCK_RATE, FLOATING_POINT_DIGIT, INITIAL_RECENT_CHAINS_GENESISHASH, PROFILE_COLORS, RELAY_CHAINS_GENESISHASH, SHORT_ADDRESS_CHARACTERS, WESTEND_GENESIS_HASH } from './constants';
 
 interface Meta {
   docs: Text[];
@@ -552,13 +553,17 @@ export const decodeHexValues = (obj: unknown) => {
     return obj;
   }
 
-  const objAsRecord = { ...obj } as Record<string, unknown>;
+  const objAsRecord = { ...obj } as Record<string, any>;
 
   Object.keys(objAsRecord).forEach((key) => {
     if (typeof objAsRecord[key] === 'string' && objAsRecord[key].startsWith('0x')) {
       objAsRecord[key] = hexToString(objAsRecord[key]);
     }
   });
+
+  if ('interior' in objAsRecord && 'x1' in objAsRecord['interior']) {
+    objAsRecord['interior'].x1 = [objAsRecord['interior'].x1];
+  }
 
   return objAsRecord;
 };
@@ -577,4 +582,28 @@ export const decodeMultiLocation = (hexString: HexString) => {
   }
 
   return decodeHexValues(decodedMultiLocation);
+};
+
+export const addressToChain = (address: string) => {
+  if (!isValidAddress(address)) {
+    console.log('Not a valid address');
+
+    return null;
+  }
+
+  if (getSubstrateAddress(address) === address) {
+    return {
+      chainName: 'Westend',
+      genesisHash: WESTEND_GENESIS_HASH
+    };
+  }
+
+  const publicKey = decodeAddress(address, true);
+
+  const chain = allChains.find(({ ss58Format }) => encodeAddress(publicKey, ss58Format) === address);
+
+  return {
+    chainName: chain?.chain,
+    genesisHash: chain?.genesisHash
+  };
 };